refactor validation part
